app/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
mysql-connector-python==8.0.32
pydantic==1.10.6
python-dotenv==1.0.0
